requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.2.0
datasets>=2.12.0
evaluate>=0.4.0
wandb>=0.15.0
spacy>=3.5.0
nltk>=3.8.1
textblob>=0.17.1
polyglot>=16.7.4
langdetect>=1.0.9
networkx>=3.1
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordnet>=0.1.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0 